+++ backend/src/models/index.js
@@ -91,17 +91,19 @@ const Dependent = require('./Dependent')(sequelize);
 const Transaction = require('./Transaction')(sequelize);
 const MemberPayment = require('./MemberPayment')(sequelize);
 const Donation = require('./Donation')(sequelize);
+const Outreach = require('./Outreach')(sequelize);
 const SmsLog = require('./SmsLog')(sequelize);
 const Group = require('./Group')(sequelize);
 const MemberGroup = require('./MemberGroup')(sequelize);
 const ZelleMemoMatch = require('./ZelleMemoMatch')(sequelize);
 
 // Define associations
-Member.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch });
-Dependent.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch });
-Transaction.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch });
-MemberPayment.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch });
-Donation.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch });
+Member.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch, Outreach });
+Dependent.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch, Outreach });
+Transaction.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch, Outreach });
+MemberPayment.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch, Outreach });
+Donation.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch, Outreach });
+Outreach.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch, Outreach });
 if (typeof Group.associate === 'function') {
   Group.associate({ Dependent, Member, Transaction, MemberPayment, Donation, SmsLog, Group, MemberGroup, ZelleMemoMatch });
 }
@@ -120,6 +122,7 @@ module.exports = {
   Transaction,
   MemberPayment,
   Donation,
+  Outreach,
   SmsLog,
   Group,
   MemberGroup,